requirements.txt
aiosqlite==0.17.0
anyio==3.6.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.0
fastapi==0.88.0
h11==0.14.0
idna==3.4
importlib-metadata==5.2.0
iso8601==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
PyJWT==1.7.1
pypika-tortoise==0.1.6
python-decouple==3.7
python-multipart==0.0.5
pytz==2022.7
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tortoise-orm==0.19.2
typing_extensions==4.4.0
uvicorn==0.20.0
zipp==3.11.0